+++ R/survival_advanced.R
@@ -22,6 +22,8 @@
 #' Time is internally calculated in days and additionally expressed in years
 #' (365.25 days). Individuals who die before follow-up start are excluded.
 #'
+#' Censoring affects inviduals date of death, immigration date and diagnosis dates.
+#'
 #' @param exposure_diagnoses A data frame containing exposure diagnoses.
 #'   Must include columns \code{ID} and \code{DATE}.
 #'
@@ -49,9 +51,11 @@
 #' \describe{
 #'   \item{plot_days}{\code{ggplot} object of cumulative incidence (days).}
 #'   \item{plot_years}{\code{ggplot} object of cumulative incidence (years).}
+#'   \item{plot_mortality}{\code{ggplot} object of cumulative incidence of death (years).}
 #'   \item{CR_days}{\code{cuminc} object with time in days.}
 #'   \item{CR_years}{\code{cuminc} object with time in years.}
-#'   \item{dsurv}{Final individual-level survival data used in the analysis.}
+#'   \item{dmodel}{Final individual-level survival data used in the main analysis.}
+#'   \item{dmortality}{Final individual-level survival data used in the death analysis.}
 #' }
 #'
 #' @seealso
@@ -89,7 +93,7 @@ survival_analysis <- function(exposure_diagnoses,
   # )
 
   internal_function <- function() {
-    .safe_inc_progress(1/3)
+    .safe_inc_progress(1/6)
 
     ## Colors for groups
     colors_groups <- c(
@@ -198,6 +202,8 @@ survival_analysis <- function(exposure_diagnoses,
       }
     }
 
+    .safe_inc_progress(2/6)
+
 
     ## PHASE 2: CALCULATE TIMES -----------------
     if(TRUE){
@@ -236,6 +242,8 @@ survival_analysis <- function(exposure_diagnoses,
         ## TODO lasketaan ajat päivinä. Loppuvaiheessa voidaan muuttaa vuosiksi
       }
     }
+
+    .safe_inc_progress(3/6)
 
 
     ## PHASE 3A: NORMAL-------------------
@@ -424,6 +432,8 @@ survival_analysis <- function(exposure_diagnoses,
       }
     }
 
+    .safe_inc_progress(4/6)
+
 
     ## PHASE 3B: MORTALITY -------------------
     ### FILTER EVENTS, CODING & RESULTS -------
@@ -506,6 +516,7 @@ survival_analysis <- function(exposure_diagnoses,
       )
     }
 
+    .safe_inc_progress(5/6)
 
     ## PHASE 6 COLLECT RESULTS ------
     if(TRUE){
@@ -519,6 +530,7 @@ survival_analysis <- function(exposure_diagnoses,
                 dmortality = dphase4b)
       .safe_inc_progress(3/3)
     }
+    .safe_inc_progress(6/6)
     return(d)
   }